Refactor to "History" to match requirements verbiage
